+++ src/pages/GodboltPage/components/index.ts
@@ -1,0 +1,3 @@
+export {default as CompileButton} from "./CompileButton"
+export {default as ShareButton} from "./ShareButton"
+export {default as SettingsDropdown} from "./SettingsDropdown"
